src/requirements.txt
Pillow==8.0.1
tensorflow==2.3.0
keras==2.4.3
numpy==1.16.4
matplotlib==3.3.3
